# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "JOIN API",
    "description": "API мини-приложения **JOIN** в MAX: афиша города, события по Пушкинской карте и поиск компании, чтобы пойти вместе.\n\n### Основной сценарий\n1. **Вход** — `POST /api/auth/login` (или регистрация) возвращает `token`.\n2. **Афиша** — `GET /api/events?city=Казань`, фильтр `pushkinCard=true`.\n3. **Сохранить событие** — `POST /api/events/{id}/like`. Если его уже сохранил подходящий человек, в ответе будет совпадение.\n4. **Позвать пойти вместе** — `POST /api/matches/{id}/request`, второй отвечает `POST /api/matches/{id}/accept`. Только после этого открывается чат.\n5. **Чат** — `GET` / `POST /api/chats/{chatId}/messages`.\n\n### Как попробовать\nПолучите токен через `POST /api/auth/login`, нажмите **Authorize** и вставьте его в `bearerAuth`. Внутри MAX вместо токена приложение передаёт заголовок `X-Max-Init-Data`.\n\n### Ответы\n* Без авторизации защищённые методы отвечают `403` с пустым телом.\n* Ошибки — JSON `ErrorResponse`: `{\"error\": \"текст для пользователя\", \"code\": \"…\"}`.\n* Списки с `page`, `size`, `sort` возвращают страницу: `content`, `totalElements`, `number`, `size`.\n* Даты — `YYYY-MM-DD`, время — `HH:mm:ss`, местное время города события.\n\nОбязательные проверки основного сценария — `DATA-API.yaml` в репозитории.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://2-27-20-14.sslip.io",
      "description": "Демо-стенд"
    },
    {
      "url": "/",
...
    }
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "maxInitData": []
    }
  ],
...
  "paths": {
    "/api/users/me/search-criteria": {
      "get": {
        "tags": [
          "Профиль"
        ],
        "summary": "Настройки поиска компании",
        "description": "Пол, возраст и вуз напарника; пустое поле — без ограничения.",
        "operationId": "getSearchCriteria",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchCriteriaResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Профиль"
        ],
        "summary": "Изменить настройки поиска компании",
        "description": "Учитываются при следующих совпадениях, причём с обеих сторон.",
        "operationId": "updateSearchCriteria",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchCriteriaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/support/ticket/read": {
      "put": {
        "tags": [
          "Поддержка"
        ],
        "summary": "Отметить ответы поддержки прочитанными",
        "operationId": "markRead",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/notifications/{id}/read": {
      "post": {
        "tags": [
          "Уведомления"
        ],
        "summary": "Отметить уведомление прочитанным",
        "operationId": "markRead_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/notifications/read-all": {
      "post": {
        "tags": [
          "Уведомления"
        ],
        "summary": "Прочитать все",
        "operationId": "markAllRead",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/matches/{id}/request": {
      "post": {
        "tags": [
          "Совпадения и приглашения"
        ],
        "summary": "Позвать пойти вместе",
        "description": "Второй получит приглашение в приложении и сообщение бота с кнопками «Пойдём» / «Не в этот раз». Чата ещё нет.",
        "operationId": "request",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Второй уже ответил на приглашение",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications": {
      "get": {
        "tags": [
          "Уведомления"
        ],
        "summary": "Уведомления",
        "description": "Новые совпадения, приглашения, сообщения — то же, что присылает бот.",
        "operationId": "getNotifications",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageNotificationResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/notifications/unread-count": {
      "get": {
        "tags": [
          "Уведомления"
        ],
        "summary": "Число непрочитанных",
        "operationId": "getUnreadCount",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Что пошло не так",
            "example": "Сегодня можно лайкнуть не больше 10 событий — возвращайся завтра"
          },
          "message": {
            "type": "string",
            "description": "То же, что `error` (есть у ошибок действий пользователя, код 409)"
          },
          "code": {
            "type": "string",
            "description": "Машинный код ошибки, если есть",
            "example": "VALIDATION_FAILED"
          }
        },
        "description": "Ошибка. Текст в `error` можно показывать пользователю"
      },
      "SearchCriteriaRequest": {
        "type": "object",
        "properties": {
          "preferredAgeMin": {
            "type": "integer",
            "format": "int32"
          },
          "preferredAgeMax": {
            "type": "integer",
            "format": "int32"
          },
          "preferredGender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE"
            ]
          },
          "preferredUniversityId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "MatchResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id совпадения",
            "format": "int64",
            "example": 30
          },
          "companionId": {
            "type": "integer",
            "description": "Id напарника",
            "format": "int64",
            "example": 40
          },
          "companionName": {
            "type": "string",
            "description": "Имя напарника",
            "example": "Боря"
          },
          "companionPhoto": {
            "type": "string",
            "description": "Фото напарника"
          },
          "companionAge": {
            "type": "integer",
            "description": "Возраст напарника",
            "format": "int32",
            "example": 21
          },
          "eventId": {
            "type": "integer",
            "description": "Id события",
            "format": "int64",
            "example": 9835
          },
          "eventTitle": {
            "type": "string",
            "description": "Событие",
            "example": "Щелкунчик"
          },
          "eventDate": {
            "type": "string",
            "description": "Дата события",
            "format": "date",
            "example": "2026-10-12"
          },
          "status": {
            "type": "string",
            "description": "NEW — никто не позвал, REQUESTED — приглашение отправлено, ACCEPTED — договорились, чат открыт, DECLINED — отказ",
            "example": "REQUESTED"
          },
          "requestedByMe": {
            "type": "boolean",
            "description": "Приглашение отправил текущий пользователь"
          },
          "chatId": {
            "type": "integer",
            "description": "Id чата — после ACCEPTED",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "description": "Когда нашлось совпадение",
            "format": "date-time"
          }
        },
        "description": "Найденный напарник по событию"
      },
...
      "SearchCriteriaResponse": {
        "type": "object",
        "properties": {
          "preferredAgeMin": {
            "type": "integer",
            "format": "int32"
          },
          "preferredAgeMax": {
            "type": "integer",
            "format": "int32"
          },
          "preferredGender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE"
            ]
          },
          "preferredUniversityId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "paged": {
            "type": "boolean"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "direction": {
            "type": "string"
          },
          "nullHandling": {
            "type": "string"
          },
          "ascending": {
            "type": "boolean"
          },
          "property": {
            "type": "string"
          },
          "ignoreCase": {
            "type": "boolean"
          }
        }
      },
...
      "NotificationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string",
            "enum": [
              "MATCH",
              "GROUP_INVITE"
            ]
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "matchId": {
            "type": "integer",
            "format": "int64"
          },
          "companionId": {
            "type": "integer",
            "format": "int64"
          },
          "companionFirstName": {
            "type": "string"
          },
          "companionPhoto": {
            "type": "string"
          },
          "eventId": {
            "type": "integer",
            "format": "int64"
          },
          "eventTitle": {
            "type": "string"
          },
          "chatId": {
            "type": "integer",
            "format": "int64"
          },
          "groupId": {
            "type": "integer",
            "format": "int64"
          },
          "read": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PageNotificationResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Некорректные данные запроса",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "JOIN доступен с 14 лет",
              "code": "VALIDATION_FAILED"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Нет авторизации (тело пустое) или нет доступа к объекту"
      },
      "NotFound": {
        "description": "Объект не найден или недоступен",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Event not found with id: 42"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Токен из POST /api/auth/login или /api/auth/register",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "maxInitData": {
        "type": "apiKey",
        "description": "window.WebApp.initData из MAX Bridge; подпись проверяется токеном бота",
        "name": "X-Max-Init-Data",
        "in": "header"
      }
    }
  }
}